refactor to introduce connector package
